Remove inactive rust-crypto dependency
This crate has known vulnerabilities and is not maintained; so let us
switch to the toolset provided by the RustCrypto GitHub org, as advised
on: https://rustsec.org/advisories/RUSTSEC-2016-0005.html

It means that I go back to a classic equality check for the hash
validation, because the fixed_time_eq() function was provided by the
unmaintained crate.
I did not found a convincing, straightforward replacement yet; so as it
is not critical for this command line tool, I will start like this.